refactor: 관련 패키지 common 디렉터리로 이동
