+++ src/main/java/im/toduck/domain/routine/domain/service/RoutineService.java
@@ -3,7 +3,7 @@ package im.toduck.domain.routine.domain.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import im.toduck.domain.routine.mapper.RoutineMapper;
+import im.toduck.domain.routine.common.mapper.RoutineMapper;
 import im.toduck.domain.routine.persistence.entity.Routine;
 import im.toduck.domain.routine.persistence.repository.RoutineRepository;
 import im.toduck.domain.routine.presentation.dto.request.RoutineCreateRequest;
