+++ src/main/java/im/toduck/domain/routine/persistence/entity/Routine.java
@@ -3,7 +3,7 @@ package im.toduck.domain.routine.persistence.entity;
 import java.time.LocalTime;
 
 import im.toduck.domain.person.PlanCategory;
-import im.toduck.domain.routine.converter.DaysOfWeekBitmaskConverter;
+import im.toduck.domain.routine.common.converter.DaysOfWeekBitmaskConverter;
 import im.toduck.domain.routine.persistence.vo.PlanCategoryColor;
 import im.toduck.domain.routine.persistence.vo.RoutineMemo;
 import im.toduck.domain.user.persistence.entity.User;
